pyproject.toml
[project]
name = "bluecompany"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Taylor Monacelli", email = "taylormonacelli@gmail.com" }
]
dependencies = [
    "mistletoe",
    "jinja2>=3.1.3",
    "platformdirs>=2.4.0",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
bluecompany = "bluecompany:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.2.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bluecompany"]
